+++ account_bank_statement_import_adyen/models/__init__.py
@@ -1,2 +1,3 @@
 from . import account_bank_statement_import
+from . import account_bank_statement_import_adyen_parser
 from . import account_journal
